fix(web): simplify program controls
