+++ .vercelignore
@@ -18,6 +18,8 @@ examples/
 .github/
 
 # Local / dev
+.env
+.env.*
 .env*.local
 .DS_Store
 keypairs/
